refactor: decompose useGameProgressController and MatchQuickActionsModal
Two god files, both pure moves with zero logic edits: no file among
the extracted modules exceeds 400 lines, all pre-existing tests pass
unmodified, and manual browser verification confirms the full
end-to-end chain (goal increment -> toast notification -> score
animation) still works correctly.

hooks/useGameProgressController.ts (603 -> 260 lines):
- hooks/gameProgress/uiReducer.ts: GameProgressUiState/Action,
  gameProgressUiReducer
- hooks/gameProgress/toastQueueReducer.ts: ToastQueueState/Action,
  QueuedToastData, toastQueueReducer
- hooks/gameProgress/matchMigration.ts: migrateLegacyMatch
- hooks/gameProgress/goalScoring.ts: LastGoalInfo,
  GoalScoreUpdateResult, calculateToastScoreDisplay,
  applyHomeGoalUpdate, applyAwayGoalUpdate, updateMatchForGoal
- hooks/useGoalToastQueue.ts: the toast subsystem as its own hook
  (queue state, the show-next-toast effect, and assembling a toast
  payload from a goal event) — takes getPlayersWhoDrink as a
  parameter since that depends on store data outside its concern
- The controller keeps an identical public return shape and now
  just orchestrates: goal/drink handlers, UI dispatch, live-score
  polling, and composing the extracted pieces

components/gameProgress/MatchQuickActionsModal.tsx (1424 lines) ->
components/gameProgress/MatchQuickActionsModal/ folder, following
the sibling MatchesGrid/ pattern:
- index.tsx: the modal shell (~340 lines)
- StatProgressBar.tsx, PossessionCircle.tsx: extracted sub-components
- styles.ts: the ~370-line createStyles
- types.ts: MatchQuickActionsModalProps
- hooks/useMatchQuickActionsAnimations.ts: the 8 Animated.Value refs
  plus the entry/goal-pop animation effects and animateButtonPress
- utils/svgArc.ts: polarToCartesian/describeArc (pure SVG geometry),
  with 11 new unit tests
folder index.tsx preserves the existing import path — no consumer
(app/gameProgress.tsx, components/index.ts) needed changes

Preserved two pre-existing type errors verbatim rather than fixing
them in passing: StatProgressBar/PossessionCircle call createStyles
with only 1 of its 5 params (harmless at runtime — the properties
they read don't depend on the other params), now at their new file
locations instead of the old MatchQuickActionsModal.tsx lines. This
phase is a pure move; unrelated pre-existing issues stay untouched.

Verified: full suite 65 suites/299 tests green (was 64/288 — +1
suite/+11 tests for svgArc). Lint 0 errors, 298 warnings (unchanged
from Phase 6 baseline). tsc exactly 130 errors (unchanged from
Phase 6 baseline) — confirmed the 2 relocated createStyles errors
via diff, nothing new. Manually seeded a game via localStorage,
opened the quick-actions modal in-browser, incremented a goal, and
confirmed the toast fired correctly ("Arsenal 1-1 Chelsea" /
"Alice, Bob DRINK!") with no console errors — proves the full
decomposed chain works end-to-end. Local `npm run test:e2e` hit the
already-documented pre-existing sandbox flakiness from the Phase 2
investigation (uniform ~29s timeouts across unrelated scenarios
like History and Legacy Import, not just gameplay ones) — not a
regression from this change.

Part of specs/019-harden-codebase-foundations (Phase 7 of 12).
