+++ components/gameProgress/MatchQuickActionsModal/index.tsx
@@ -1,325 +1,31 @@
 /**
- * @file MatchQuickActionsModal.tsx
- * @description Modal + sub-components for quick match actions and stat visualizations (progress bars, possession circle) for a selected match.
+ * @file MatchQuickActionsModal/index.tsx
+ * @description Modal for quick match actions and stat visualizations (progress bars, possession circle) for a selected match.
  */
 
-import React, { useRef, useEffect, useMemo, useState } from "react";
+import React, { useMemo, useState } from "react";
 import {
   View,
   Text,
   TouchableOpacity,
   Modal,
   ScrollView,
-  StyleSheet,
   SafeAreaView,
   Animated,
   Image,
   useWindowDimensions,
 } from "react-native";
-import Svg, { Path, Rect } from "react-native-svg";
-import { Match, Player } from "../../store/store";
 import { Ionicons } from "@expo/vector-icons";
-import { useTeamLogo } from "../../hooks/useTeamLogo";
-import { MatchWithScore } from "../../types/matchScores";
-import { useColors } from "../../app/style/theme";
-
-/** Props for MatchQuickActionsModal. */
-interface MatchQuickActionsModalProps {
-  /** Whether the modal is visible. */
-  isVisible: boolean;
-  /** Callback invoked to close the modal. */
-  onClose: () => void;
-  /** ID of the currently selected match for quick actions; null if none. */
-  selectedMatchId: string | null;
-  /** All match objects (from the store). */
-  matches: Match[];
-  /** All player objects (from the store). */
-  players: Player[];
-  /** The ID of the designated common match. */
-  commonMatchId: string;
-  /** Mapping of player IDs to the match IDs they are assigned to. */
-  playerAssignments: Record<string, string[]>;
-  /** Increment a team's goal count for a match. */
-  handleGoalIncrement: (matchId: string, team: "home" | "away") => void;
-  /** Decrement a team's goal count for a match. */
-  handleGoalDecrement: (matchId: string, team: "home" | "away") => void;
-  /** Live match data (if available) used to render API-controlled matches. */
-  liveMatches: MatchWithScore[];
-}
+import { Player } from "../../../store/store";
+import { useTeamLogo } from "../../../hooks/useTeamLogo";
+import { useMatchQuickActionsAnimations } from "../../../hooks/useMatchQuickActionsAnimations";
+import { useColors } from "../../../app/style/theme";
+import { StatProgressBar } from "./StatProgressBar";
+import { PossessionCircle } from "./PossessionCircle";
+import { createStyles } from "./styles";
+import { MatchQuickActionsModalProps } from "./types";
 
 const useThemed = () => useColors();
-
-/**
- * Horizontal bar comparison for a stat (home vs away).
- * @param {{homeValue:number, awayValue:number, label:string, isPercentage?:boolean}} props Component props.
- * @returns {JSX.Element} Rendered stat progress bar.
- */
-const StatProgressBar = ({
-  homeValue,
-  awayValue,
-  label,
-  isPercentage = false,
-}: {
-  homeValue: number;
-  awayValue: number;
-  label: string;
-  isPercentage?: boolean;
-}) => {
-  const colors = useThemed();
-  const styles = useMemo(() => createStyles(colors), [colors]);
-  const maxValue = Math.max(1, homeValue, awayValue); // Ensure maxValue is at least 1 to avoid division by zero
-  const homePercent = (homeValue / maxValue) * 100;
-  const awayPercent = (awayValue / maxValue) * 100;
-
-  const barHeight = 8; // Height of the progress bars
-  const totalBarWidth = 100; // Represents 100% width for calculations
-  const homeBarWidth = (homePercent / 100) * (totalBarWidth / 2);
-  const awayBarWidth = (awayPercent / 100) * (totalBarWidth / 2);
-
-  const homeColor = colors.primary;
-  const awayColor = colors.awayTeam;
-  const dividerColor = colors.darkSurface;
-
-  return (
-    <View
-      style={styles.statProgressContainer}
-      accessible={true}
-      accessibilityLabel={`${label}: Home ${homeValue}${
-        isPercentage ? "%" : ""
-      }, Away ${awayValue}${isPercentage ? "%" : ""}`}
-    >
-      <Text style={styles.statValue}>
-        {homeValue}
-        {isPercentage ? "%" : ""}
-      </Text>
-
-      <View style={styles.statProgressWrapper}>
-        <Text style={styles.statProgressLabel}>{label}</Text>
-        <View style={styles.svgProgressBarContainer}>
-          <Svg height={barHeight} width="100%">
-            {/* Background for Home Side */}
-            <Rect
-              x="0"
-              y="0"
-              width="50%"
-              height={barHeight}
-              fill={styles.homeProgressArea.backgroundColor}
-              rx={styles.homeProgressBar.borderTopLeftRadius} // Optional: for rounded corners
-              ry={styles.homeProgressBar.borderTopLeftRadius}
-            />
-            {/* Home Progress */}
-            <Rect
-              x={`${50 - homeBarWidth}%`} // Start from the right edge of the home area and draw left
-              y="0"
-              width={`${homeBarWidth}%`}
-              height={barHeight}
-              fill={homeColor}
-              rx={styles.homeProgressBar.borderTopLeftRadius}
-              ry={styles.homeProgressBar.borderTopLeftRadius}
-            />
-
-            {/* Background for Away Side */}
-            <Rect
-              x="50%"
-              y="0"
-              width="50%"
-              height={barHeight}
-              fill={styles.awayProgressArea.backgroundColor}
-              rx={styles.awayProgressBar.borderTopRightRadius} // Optional: for rounded corners
-              ry={styles.awayProgressBar.borderTopRightRadius}
-            />
-            {/* Away Progress */}
-            <Rect
-              x="50%"
-              y="0"
-              width={`${awayBarWidth}%`}
-              height={barHeight}
-              fill={awayColor}
-              rx={styles.awayProgressBar.borderTopRightRadius}
-              ry={styles.awayProgressBar.borderTopRightRadius}
-            />
-
-            {/* Center Divider */}
-            <Rect
-              x="50%"
-              y="0"
-              width={styles.progressDivider.width}
-              height={barHeight}
-              fill={dividerColor}
-              transform="translate(-1)" // Adjust to center the 2px divider
-            />
-          </Svg>
-        </View>
-      </View>
-
-      <Text style={styles.statValue}>
-        {awayValue}
-        {isPercentage ? "%" : ""}
-      </Text>
-    </View>
-  );
-};
-
-// Helper functions for SVG arc
-/**
- * Convert polar coordinates to cartesian for an SVG arc.
- * @param {number} centerX X coordinate of center.
- * @param {number} centerY Y coordinate of center.
- * @param {number} radius Arc radius.
- * @param {number} angleInDegrees Angle in degrees.
- * @returns {{x:number,y:number}} Cartesian coordinates.
- */
-function polarToCartesian(
-  centerX: number,
-  centerY: number,
-  radius: number,
-  angleInDegrees: number
-) {
-  const angleInRadians = ((angleInDegrees - 90) * Math.PI) / 180;
-  return {
-    x: centerX + radius * Math.cos(angleInRadians),
-    y: centerY + radius * Math.sin(angleInRadians),
-  };
-}
-/**
- * Describe an SVG arc path segment.
- * Handles full-circle edge cases by slightly shrinking span.
- * @param {number} x Center X.
- * @param {number} y Center Y.
- * @param {number} radius Arc radius.
- * @param {number} startAngle Starting angle (deg).
- * @param {number} endAngle Ending angle (deg).
- * @param {boolean} [counterClockwise=false] Direction flag; true = CCW.
- * @returns {string} SVG path definition or empty string if no arc span.
- */
-function describeArc(
-  x: number,
-  y: number,
-  radius: number,
-  startAngle: number,
-  endAngle: number,
-  counterClockwise: boolean = false
-): string {
-  if (startAngle === endAngle || Math.abs(startAngle - endAngle) < 0.01) {
-    return ""; // No arc to draw
-  }
-  // Ensure endAngle is slightly different from startAngle if it's a full circle from 0 to 360
-  if (
-    Math.abs(endAngle - startAngle - 360) < 0.01 ||
-    Math.abs(endAngle - startAngle + 360) < 0.01
-  ) {
-    endAngle = startAngle + (endAngle > startAngle ? 359.99 : -359.99);
-  } else if (endAngle >= 360 && startAngle === 0) {
-    endAngle = 359.99;
-  }
-
-  const start = polarToCartesian(x, y, radius, endAngle);
-  const end = polarToCartesian(x, y, radius, startAngle);
-
-  const largeArcFlag = Math.abs(endAngle - startAngle) <= 180 ? "0" : "1";
-  const sweepFlag = counterClockwise ? "1" : "0"; // 0 for clockwise, 1 for counter-clockwise
-
-  const d = [
-    "M",
-    start.x,
-    start.y,
-    "A",
-    radius,
-    radius,
-    0,
-    largeArcFlag,
-    sweepFlag,
-    end.x,
-    end.y,
-  ].join(" ");
-
-  return d;
-}
-
-/**
- * Doughnut possession chart (home vs away) rendered with two opposing arc paths.
- * @param {{homeValue:number, awayValue:number}} props Component props.
- * @returns {JSX.Element} Possession visualization.
- */
-const PossessionCircle = ({
-  homeValue,
-  awayValue,
-}: {
-  homeValue: number;
-  awayValue: number;
-}) => {
-  const colors = useThemed();
-  const styles = useMemo(() => createStyles(colors), [colors]);
-  const total = homeValue + awayValue;
-  const normalizedHome = total > 0 ? Math.round((homeValue / total) * 100) : 0;
-  const normalizedAway = total > 0 ? Math.round((awayValue / total) * 100) : 0;
-
-  const size = 90; // Diameter of the doughnut
-  const strokeWidth = 18; // Thickness of the doughnut ring
-  const radius = (size - strokeWidth) / 2;
-  const cx = size / 2;
-  const cy = size / 2;
-
-  const homeColor = colors.primary;
-  const awayColor = colors.awayTeam;
-  const trackColor = colors.borderLight; // Background track color
-
-  const homeAngle = (normalizedHome / 100) * 360;
-  const awayAngle = (normalizedAway / 100) * 360;
-
-  // Path for the background track (full circle)
-  const trackPath = describeArc(cx, cy, radius, 0, 359.99);
-
-  // Path for home team's possession - counter-clockwise from 0
-  const homeArcPath =
-    homeAngle > 0
-      ? describeArc(cx, cy, radius, 0, -homeAngle, true) // Negative angle for counter-clockwise
-      : "";
-
-  // Path for away team's possession - clockwise from 0
-  const awayArcPath =
-    awayAngle > 0 ? describeArc(cx, cy, radius, 0, awayAngle) : "";
-
-  return (
-    <View style={styles.possessionContainer}>
-      <Text style={styles.statValue}>{normalizedHome}%</Text>
-
-      <View style={styles.possessionCircleContainer}>
-        <Text style={styles.statProgressLabel}>Possession</Text>
-        <View style={styles.circleWrapper}>
-          <Svg width={size} height={size} viewBox={`0 0 ${size} ${size}`}>
-            <Path
-              d={trackPath}
-              stroke={trackColor}
-              strokeWidth={strokeWidth}
-              fill="none"
-            />
-            {homeArcPath ? (
-              <Path
-                d={homeArcPath}
-                stroke={homeColor}
-                strokeWidth={strokeWidth}
-                fill="none"
-                strokeLinecap="round"
-              />
-            ) : null}
-            {awayArcPath ? (
-              <Path
-                d={awayArcPath}
-                stroke={awayColor}
-                strokeWidth={strokeWidth}
-                fill="none"
-                strokeLinecap="round"
-              />
-            ) : null}
-          </Svg>
-        </View>
-      </View>
-
-      <Text style={styles.statValue}>{normalizedAway}%</Text>
-    </View>
-  );
-};
 
 /** Quick actions & stats modal for selected match (scores, scorers, assignments, stats). */
 const MatchQuickActionsModal: React.FC<MatchQuickActionsModalProps> = ({
@@ -358,16 +64,6 @@ const MatchQuickActionsModal: React.FC<MatchQuickActionsModalProps> = ({
     [colors, isWideLayout, modalWidth, playerColumnCount, screenHeight],
   );
   const [activeTab, setActiveTab] = useState("overview");
-
-  // Animation values for UI elements
-  const closeButtonAnim = useRef(new Animated.Value(1)).current;
-  const goalValueAnimHome = useRef(new Animated.Value(1)).current;
-  const goalValueAnimAway = useRef(new Animated.Value(1)).current;
-  const modalContentAnim = useRef(new Animated.Value(0)).current;
-  const incrementAnimHome = useRef(new Animated.Value(1)).current;
-  const decrementAnimHome = useRef(new Animated.Value(1)).current;
-  const incrementAnimAway = useRef(new Animated.Value(1)).current;
-  const decrementAnimAway = useRef(new Animated.Value(1)).current;
 
   /** Selected match entity (or null). */
   const match = useMemo(() => {
@@ -408,9 +104,22 @@ const MatchQuickActionsModal: React.FC<MatchQuickActionsModalProps> = ({
     );
   }, [liveMatchData]);
 
-  // References to store previous goal values for animation triggering
-  const prevGoalsHomeRef = useRef(match?.homeGoals ?? 0);
-  const prevGoalsAwayRef = useRef(match?.awayGoals ?? 0);
+  const {
+    closeButtonAnim,
+    goalValueAnimHome,
+    goalValueAnimAway,
+    modalContentAnim,
+    incrementAnimHome,
+    decrementAnimHome,
+    incrementAnimAway,
+    decrementAnimAway,
+    animateButtonPress,
+  } = useMatchQuickActionsAnimations({
+    isVisible,
+    match,
+    liveMatchData,
+    isApiControlledMatch,
+  });
 
   /** Players impacted by this match (assigned or common). */
   const affectedPlayers = useMemo(() => {
@@ -424,107 +133,6 @@ const MatchQuickActionsModal: React.FC<MatchQuickActionsModalProps> = ({
 
   // Check if the selected match is the common match
   const isCommonMatch = match ? match.id === commonMatchId : false;
-
-  /**
-   * Run press animation on a control button.
-   * @param {Animated.Value} buttonAnim Animated value controlling scale.
-   */
-  const animateButtonPress = (buttonAnim: Animated.Value) => {
-    Animated.sequence([
-      Animated.timing(buttonAnim, {
-        toValue: 0.9,
-        duration: 100,
-        useNativeDriver: true,
-      }),
-      Animated.timing(buttonAnim, {
-        toValue: 1,
-        duration: 150,
-        useNativeDriver: true,
-      }),
-    ]).start();
-  };
-
-  /** Entry animation when modal becomes visible. */
-  useEffect(() => {
-    if (isVisible && match) {
-      Animated.timing(modalContentAnim, {
-        toValue: 1,
-        duration: 200,
-        useNativeDriver: true,
-      }).start();
-    }
-
-    // Reset animation when closing
-    return () => {
-      if (!isVisible) {
-        modalContentAnim.setValue(0);
-      }
-    };
-  }, [isVisible, match, modalContentAnim]);
-
-  /** Animate home goals when the displayed value changes. */
-  useEffect(() => {
-    // Get the current score (prioritize live data if available)
-    const currentScore = isApiControlledMatch
-      ? liveMatchData?.homeScore ?? 0
-      : match?.homeGoals ?? 0;
-
-    if (prevGoalsHomeRef.current !== currentScore) {
-      // Run animation
-      Animated.sequence([
-        Animated.timing(goalValueAnimHome, {
-          toValue: 1.3,
-          duration: 200,
-          useNativeDriver: true,
-        }),
-        Animated.timing(goalValueAnimHome, {
-          toValue: 1,
-          duration: 200,
-          useNativeDriver: true,
-        }),
-      ]).start();
-
-      prevGoalsHomeRef.current = currentScore;
-    }
-  }, [
-    match?.homeGoals,
-    liveMatchData?.homeScore,
-    isApiControlledMatch,
-    goalValueAnimHome,
-    liveMatchData,
-  ]); // Added goalValueAnimHome and liveMatchData to dependency array
-
-  /** Animate away goals when the displayed value changes. */
-  useEffect(() => {
-    // Get the current score (prioritize live data if available)
-    const currentScore = isApiControlledMatch
-      ? liveMatchData?.awayScore ?? 0
-      : match?.awayGoals ?? 0;
-
-    if (prevGoalsAwayRef.current !== currentScore) {
-      // Run animation
-      Animated.sequence([
-        Animated.timing(goalValueAnimAway, {
-          toValue: 1.3,
-          duration: 200,
-          useNativeDriver: true,
-        }),
-        Animated.timing(goalValueAnimAway, {
-          toValue: 1,
-          duration: 200,
-          useNativeDriver: true,
-        }),
-      ]).start();
-
-      prevGoalsAwayRef.current = currentScore;
-    }
-  }, [
-    match?.awayGoals,
-    liveMatchData?.awayScore,
-    isApiControlledMatch,
-    goalValueAnimAway,
-    liveMatchData,
-  ]); // Added goalValueAnimAway and liveMatchData to dependency array
 
   /**
    * Distribute affected players into three columns for display.
@@ -1047,378 +655,5 @@ const MatchQuickActionsModal: React.FC<MatchQuickActionsModalProps> = ({
     </SafeAreaView>
   );
 };
-
-/** Style creator for the MatchQuickActionsModal and its sub-components. */
-const createStyles = (
-  colors: ReturnType<typeof useColors>,
-  modalWidth: number,
-  screenHeight: number,
-  isWideLayout: boolean,
-  playerColumnCount: number,
-) =>
-  StyleSheet.create({
-    overlayTouchable: {
-      position: "absolute",
-      top: 0,
-      bottom: 0,
-      left: 0,
-      right: 0,
-      backgroundColor: colors.backgroundModalOverlay,
-      zIndex: 1000,
-    },
-    centeredView: {
-      flex: 1,
-      justifyContent: "center",
-      alignItems: "center",
-      padding: isWideLayout ? 24 : 16,
-    },
-    modalContainer: {
-      width: modalWidth,
-      maxWidth: isWideLayout ? 960 : 420,
-      backgroundColor: colors.surface,
-      borderRadius: 16,
-      overflow: "hidden",
-      shadowColor: colors.black,
-      shadowOffset: { width: 0, height: 2 },
-      shadowOpacity: 0.2,
-      shadowRadius: 6,
-      elevation: 5,
-      zIndex: 1001,
-      maxHeight: Math.min(screenHeight * 0.84, isWideLayout ? 760 : 680),
-    },
-    modalInnerContainer: {
-      width: "100%",
-    },
-    scrollContent: {
-      padding: 16,
-      alignItems: "center",
-    },
-    // Match header styling to match MatchesGrid
-    matchHeaderSection: {
-      flexDirection: "row",
-      alignItems: "center",
-      justifyContent: "space-between",
-      width: "100%",
-      paddingBottom: 12,
-    },
-    matchTeamContainer: {
-      flex: 1,
-      alignItems: "center",
-      justifyContent: "center",
-      paddingHorizontal: 8,
-    },
-    matchTeamLogo: {
-      width: 50,
-      height: 50,
-      resizeMode: "contain",
-      marginBottom: 8,
-    },
-    matchTeamName: {
-      fontSize: 14,
-      fontWeight: "600",
-      textAlign: "center",
-      color: colors.textSecondary,
-      width: "100%",
-    },
-    matchVsBadge: {
-      width: 32,
-      height: 32,
-      borderRadius: 16,
-      backgroundColor: colors.background,
-      justifyContent: "center",
-      alignItems: "center",
-      marginHorizontal: 10,
-    },
-    matchVsText: {
-      fontSize: 12,
-      fontWeight: "700",
-      color: colors.textMuted,
-    },
-    // Common match badge
-    commonMatchBadge: {
-      backgroundColor: colors.success,
-      paddingHorizontal: 12,
-      paddingVertical: 4,
-      borderRadius: 12,
-      marginTop: 4,
-      marginBottom: 8,
-    },
-    commonMatchText: {
-      color: colors.white,
-      fontSize: 12,
-      fontWeight: "600",
-    },
-    divider: {
-      height: 1,
-      backgroundColor: colors.borderLighter,
-      width: "100%",
-      marginVertical: 12,
-    },
-    // Tab container
-    tabContainer: {
-      flexDirection: "row",
-      justifyContent: isWideLayout ? "center" : "space-around",
-      width: "100%",
-      marginBottom: 12,
-    },
-    tabButton: {
-      paddingVertical: 8,
-      paddingHorizontal: 16,
-      borderRadius: 8,
-      backgroundColor: colors.backgroundSubtle,
-    },
-    activeTab: {
-      backgroundColor: colors.primary,
-    },
-    tabText: {
-      fontSize: 14,
-      fontWeight: "600",
-      color: colors.textSecondary,
-    },
-    activeTabText: {
-      color: colors.white,
-    },
-    // For API matches - read only display
-    scoreContainer: {
-      flexDirection: "row",
-      alignItems: "center",
-      justifyContent: "center",
-      width: "100%",
-      paddingVertical: 16,
-    },
-    scoreValue: {
-      alignItems: "center",
-      paddingHorizontal: 12,
-    },
-    scoreText: {
-      fontSize: 42,
-      fontWeight: "bold",
-      color: colors.primary,
-    },
-    scoreSeparator: {
-      paddingHorizontal: 8,
-    },
-    scoreSeparatorText: {
-      fontSize: 36,
-      fontWeight: "bold",
-      color: colors.textMuted,
-    },
-    // For editable matches - with controls
-    goalActions: {
-      flexDirection: "row",
-      alignItems: "center",
-      justifyContent: "space-between",
-      width: "100%",
-      paddingVertical: 8,
-    },
-    teamGoalControls: {
-      flex: 1,
-      alignItems: "center",
-    },
-    teamLabel: {
-      fontSize: 14,
-      fontWeight: "600",
-      color: colors.textSecondary,
-      marginBottom: 4,
-    },
-    scoreControlRow: {
-      flexDirection: "row",
-      alignItems: "center",
-      justifyContent: "center",
-    },
-    goalCounter: {
-      alignItems: "center",
-      paddingHorizontal: 20,
-    },
-    goalValue: {
-      fontSize: 32,
-      fontWeight: "bold",
-      color: colors.primary,
-    },
-    actionButton: {
-      width: 32,
-      height: 32,
-      borderRadius: 22,
-      justifyContent: "center",
-      alignItems: "center",
-      shadowColor: colors.black,
-      shadowOffset: { width: 0, height: 2 },
-      shadowOpacity: 0.1,
-      shadowRadius: 3,
-      elevation: 2,
-    },
-    blueButton: {
-      backgroundColor: colors.primary,
-    },
-    // Goal Scorers display
-    goalScorersContainer: {
-      flexDirection: "row",
-      width: "100%",
-      minHeight: 0, // Minimum height even when empty
-      maxHeight: 120, // Maximum height before scrolling
-    },
-    teamScorersColumn: {
-      flex: 1,
-      alignItems: "center",
-      paddingHorizontal: 4,
-    },
-    scorerContainer: {
-      width: "100%",
-      alignItems: "center",
-      paddingVertical: 4,
-    },
-    scorerText: {
-      fontSize: 11,
-      color: colors.textSecondary,
-      marginVertical: 2,
-      fontStyle: "italic",
-      textAlign: "center",
-    },
-    // Players section
-    sectionHeader: {
-      flexDirection: "row",
-      alignItems: "center",
-      alignSelf: "flex-start",
-      marginBottom: 8,
-    },
-    sectionTitle: {
-      fontSize: 15,
-      fontWeight: "600",
-      color: colors.textSecondary,
-      marginLeft: 8,
-    },
-    // Player section - super compact layout
-    compactContainer: {
-      flexDirection: playerColumnCount === 1 ? "column" : "row",
-      width: "100%",
-      marginBottom: 12,
-      justifyContent: "flex-start",
-    },
-    playerColumn: {
-      flex: playerColumnCount === 1 ? undefined : 1,
-      width: playerColumnCount === 1 ? "100%" : undefined,
-      marginHorizontal: playerColumnCount === 1 ? 0 : 2,
-    },
-    compactPlayerCard: {
-      backgroundColor: colors.backgroundSubtle,
-      borderRadius: 6,
-      paddingVertical: 4, // Even more compact
-      paddingHorizontal: 6,
-      marginBottom: 4,
-      marginHorizontal: 2,
-    },
-    compactPlayerName: {
-      fontSize: 12, // Smaller font for compactness
-      fontWeight: "500",
-      color: colors.textSecondary,
-      textAlign: "center",
-    },
-    emptyStateContainer: {
-      flexDirection: "row",
-      alignItems: "center",
-      padding: 10,
-      backgroundColor: colors.backgroundLight,
-      borderRadius: 8,
-      marginBottom: 12,
-    },
-    noPlayersText: {
-      fontSize: 14,
-      color: colors.textMuted,
-      fontStyle: "italic",
-      marginLeft: 8,
-    },
-    // Close button
-    closeButton: {
-      backgroundColor: colors.backgroundSubtle,
-      paddingVertical: 10,
-      paddingHorizontal: 16,
-      borderRadius: 10,
-      alignItems: "center",
-      width: "100%",
-    },
-    closeButtonText: {
-      fontSize: 15,
-      fontWeight: "600",
-      color: colors.textSecondary,
-    },
-    // Match statistics section
-    statisticsContainer: {
-      width: "100%",
-      paddingVertical: 8,
-      paddingHorizontal: 8,
-    },
-    statProgressContainer: {
-      flexDirection: "row",
-      alignItems: "center",
-      marginVertical: 8, // Increased margin slightly
-      width: "100%",
-    },
-    statProgressWrapper: {
-      flex: 3,
-      marginHorizontal: 8, // Add some horizontal margin
-    },
-    statProgressLabel: {
-      fontSize: 12,
-      color: colors.textSecondary,
-      textAlign: "center",
-      marginBottom: 4,
-    },
-    svgProgressBarContainer: {
-      // New or repurposed style for SVG wrapper
-      flexDirection: "row",
-      alignItems: "center",
-      width: "100%",
-      height: 8, // Match barHeight
-    },
-    statValue: {
-      flex: 1,
-      textAlign: "center",
-      fontSize: 14,
-      fontWeight: "bold",
-      color: colors.textSecondary,
-    },
-    homeProgressArea: {
-      // Used for background color reference
-      backgroundColor: colors.primaryLight,
-    },
-    awayProgressArea: {
-      // Used for background color reference
-      backgroundColor: colors.playerItemOddBackground,
-    },
-    progressDivider: {
-      // Used for width and color reference
-      width: 2,
-      backgroundColor: colors.darkSurface,
-    },
-    homeProgressBar: {
-      // Used for borderRadius reference
-      borderTopLeftRadius: 3,
-      borderBottomLeftRadius: 3,
-    },
-    awayProgressBar: {
-      // Used for borderRadius reference
-      borderTopRightRadius: 3,
-      borderBottomRightRadius: 3,
-    },
-    possessionContainer: {
-      flexDirection: "row",
-      alignItems: "center",
-      justifyContent: "space-between",
-      marginVertical: 15,
-      width: "100%",
-    },
-    possessionCircleContainer: {
-      flex: 3,
-      alignItems: "center",
-    },
-    circleWrapper: {
-      width: 100,
-      height: 100,
-      alignItems: "center",
-      justifyContent: "center",
-      position: "relative",
-      marginTop: 5,
-    },
-  });
 
 export default MatchQuickActionsModal;
